+++ src/cargo/parse.rs
@@ -4,8 +4,6 @@ use std::io::{Read, Error as IoError};
 use std::borrow::Cow;
 use std::fs::File;
 use toml::{Parser, ParserError, Value};
-use chrono::UTC;
-use semver::{Version, Identifier, SemVerError};
 
 macro_rules! toml_val {
     ($toml:ident [ $key:expr ] . $cast:ident ( )) => ({
@@ -18,7 +16,6 @@ macro_rules! toml_val {
 /// The source can either be a relative filepath or a byte buffer.
 #[derive(Debug, PartialEq)]
 pub struct CargoParseArgs<'a> {
-    pub dev: bool,
     pub buf: CargoBufKind<'a>,
 }
 
@@ -47,17 +44,12 @@ pub fn parse_toml<'a>(args: CargoParseArgs<'a>) -> Result<CargoConfig, CargoPars
     let toml = parser.parse().ok_or(CargoParseError::Toml { errs: parser.errors })?;
 
     let is_dylib = is_dylib(&toml).unwrap_or(false);
-    
+
     if !is_dylib {
         Err(CargoParseError::NotADyLib)?;
     }
 
-    let mut config = parse_config_from_toml(&toml)?;
-
-    if args.dev {
-        let dev_version = get_dev_version(&config.version)?;
-        config.version = dev_version.into();
-    }
+    let config = parse_config_from_toml(&toml)?;
 
     Ok(config)
 }
@@ -129,47 +121,6 @@ fn is_dylib(toml: &BTreeMap<String, Value>) -> Result<bool, CargoParseError> {
     }
 }
 
-/// Make a dev version string.
-fn get_dev_version(ver: &str) -> Result<String, CargoVersionError> {
-    let mut ver = Version::parse(ver)?;
-    let build = UTC::now().timestamp();
-
-    if build < 0 {
-        Err(CargoVersionError::PreEpoch)?;
-    }
-
-    let build = build as u64;
-
-    add_pretag(&mut ver, "dev", build);
-
-    Ok(ver.to_string())
-}
-
-fn add_pretag(ver: &mut Version, tag: &str, num: u64) {
-    if ver.pre.len() == 0 {
-        ver.pre.push(Identifier::AlphaNumeric(tag.into()));
-    }
-
-    ver.pre.push(Identifier::Numeric(num));
-
-    ver.build = vec![];
-}
-
-quick_error!{
-    /// An error encountered while updating a semver version.
-    #[derive(Debug)]
-    pub enum CargoVersionError {
-        Parse(err: SemVerError) {
-            cause(err)
-            display("Error adding dev pretag\nCaused by: {}", err)
-            from()
-        }
-        PreEpoch {
-            display("Current timestamp is before the epoch\nYou are either a time traveller or there's an error with your clock")
-        }
-    }
-}
-
 quick_error!{
     /// An error encountered while parsing Cargo configuration.
     #[derive(Debug)]
@@ -208,51 +159,16 @@ quick_error!{
         NotADyLib {
             display("The crate must include `dylib` in `lib.crate-type`")
         }
-        /// An error parsing the version.
-        Version(err: CargoVersionError) {
-            cause(err)
-            display("Error parsing config\nCaused by: {}", err)
-            from()
-        }
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use semver::Version;
-
-    #[test]
-    fn add_pretag_and_build() {
-        let mut ver = Version::parse("0.0.1").unwrap();
-
-        add_pretag(&mut ver, "dev", 2);
-
-        assert_eq!("0.0.1-dev.2", &ver.to_string());
-    }
-
-    #[test]
-    fn use_existing_pretag() {
-        let mut ver = Version::parse("0.0.1-carrots1").unwrap();
-
-        add_pretag(&mut ver, "dev", 2);
-
-        assert_eq!("0.0.1-carrots1.2", &ver.to_string());
-    }
-
-    #[test]
-    fn use_existing_pretag_ignore_build() {
-        let mut ver = Version::parse("0.0.1-carrots+1").unwrap();
-
-        add_pretag(&mut ver, "dev", 2);
-
-        assert_eq!("0.0.1-carrots.2", &ver.to_string());
-    }
 
     #[test]
     fn parse_toml_from_file() {
-        let args = CargoParseArgs { 
-            dev: false, 
+        let args = CargoParseArgs {
             buf: CargoBufKind::FromFile { path: "tests/native/Cargo.toml".into() } 
         };
 
@@ -272,8 +188,7 @@ mod tests {
             crate-type = ["rlib", "dylib"]
         "#;
 
-        let args = CargoParseArgs { 
-            dev: false, 
+        let args = CargoParseArgs {
             buf: CargoBufKind::FromBuf { buf: toml.as_bytes().into() }
         };
 
@@ -291,8 +206,7 @@ mod tests {
 
     macro_rules! assert_inavlid {
         ($input:expr, $err:pat) => ({
-            let args = CargoParseArgs { 
-                dev: false, 
+            let args = CargoParseArgs {
                 buf: CargoBufKind::FromBuf { buf: $input.as_bytes().into() }
             };
