+++ src/main.rs
@@ -8,10 +8,6 @@ extern crate toml;
 extern crate semver;
 extern crate chrono;
 
-#[cfg(test)]
-#[macro_use]
-mod test_utils;
-
 #[macro_use]
 mod macros;
 
@@ -29,26 +25,31 @@ use clap::ArgMatches;
 fn main() {
     let args = args::app().get_matches();
 
+    // run pack command
     if let Some(args) = args.subcommand_matches(args::PACK_CMD) {
-        match build(args) {
+        match pack(args) {
             Ok(_) => {
                 println!("{}", Green.paint("The build finished successfully"));
             }
             Err(e) => {
                 println!("{}", Red.paint(e));
-                println!("\n{}",
-                         Red.bold().paint("The build did not finish successfully"));
+                println!("\n{}", Red.bold().paint("The build did not finish successfully"));
             }
         }
     }
+    // print help and exit
     else {
         args::app().print_help().unwrap();
         println!("");
     }
 }
 
-fn build(args: &ArgMatches) -> Result<(), Box<Error>> {
-    let cargo_toml = pass!("reading cargo manifest" => args => cargo::parse_toml);
+fn pack(args: &ArgMatches) -> Result<(), Box<Error>> {
+    let mut cargo_toml = pass!("reading cargo manifest" => args => cargo::parse_toml);
+
+    let local = pass!("adding local version tag" => &cargo_toml => cargo::local_version_tag);
+
+    cargo_toml.version = local.version;
 
     let cargo_lib = pass!("building Rust lib" => (args, &cargo_toml) => |args| {
         let result = cargo::build_lib(args);
